requirements.txt
click==7.1.2
requests==2.31.0
python-dotenv==1.0.1
croniter==3.0.3